Update upgrade script for new database package
