+++ database/schema.py
@@ -1,24 +1,5 @@
-# database.py - Enhanced with Objectives
-import sqlite3
-import os
-from contextlib import contextmanager
+from .connection import get_db_connection
 
-DATABASE_PATH = os.path.join('data', 'students.db')
-
-def get_db():
-    """Get database connection."""
-    conn = sqlite3.connect(DATABASE_PATH)
-    conn.row_factory = sqlite3.Row  # Enable dict-like access
-    return conn
-
-@contextmanager
-def get_db_connection():
-    """Context manager for database connections."""
-    conn = get_db()
-    try:
-        yield conn
-    finally:
-        conn.close()
 
 def init_db():
     """Initialize database with all tables."""
@@ -38,7 +19,7 @@ def init_db():
                 updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
             )
         ''')
-        
+
         # Goals table
         conn.execute('''
             CREATE TABLE IF NOT EXISTS goals (
@@ -51,7 +32,7 @@ def init_db():
                 FOREIGN KEY (student_id) REFERENCES students (id)
             )
         ''')
-        
+
         # NEW: Objectives table
         conn.execute('''
             CREATE TABLE IF NOT EXISTS objectives (
@@ -65,7 +46,7 @@ def init_db():
                 FOREIGN KEY (goal_id) REFERENCES goals (id)
             )
         ''')
-        
+
         # Sessions table
         conn.execute('''
             CREATE TABLE IF NOT EXISTS sessions (
@@ -82,7 +63,7 @@ def init_db():
                 FOREIGN KEY (student_id) REFERENCES students (id)
             )
         ''')
-        
+
         # Enhanced trial logs table - now links to objectives
         conn.execute('''
             CREATE TABLE IF NOT EXISTS trial_logs (
@@ -102,7 +83,7 @@ def init_db():
                 FOREIGN KEY (goal_id) REFERENCES goals (id)
             )
         ''')
-        
+
         # SOAP notes table
         conn.execute('''
             CREATE TABLE IF NOT EXISTS soap_notes (
@@ -117,75 +98,13 @@ def init_db():
                 FOREIGN KEY (session_id) REFERENCES sessions (id)
             )
         ''')
-        
+
         # Create indexes for better performance
         conn.execute('CREATE INDEX IF NOT EXISTS idx_goals_student ON goals(student_id)')
         conn.execute('CREATE INDEX IF NOT EXISTS idx_objectives_goal ON objectives(goal_id)')
         conn.execute('CREATE INDEX IF NOT EXISTS idx_sessions_student ON sessions(student_id)')
         conn.execute('CREATE INDEX IF NOT EXISTS idx_trials_session ON trial_logs(session_id)')
         conn.execute('CREATE INDEX IF NOT EXISTS idx_trials_objective ON trial_logs(objective_id)')
-        
+
         conn.commit()
 
-def add_sample_data():
-    """Add some sample data for testing."""
-    with get_db_connection() as conn:
-        # Check if we already have data
-        cursor = conn.execute('SELECT COUNT(*) FROM students')
-        if cursor.fetchone()[0] > 0:
-            print("Sample data already exists")
-            return
-        
-        # Add sample student
-        cursor = conn.execute('''
-            INSERT INTO students (first_name, last_name, grade_level, pronouns)
-            VALUES (?, ?, ?, ?)
-        ''', ('Alex', 'Johnson', '3rd Grade', 'they/them'))
-        student_id = cursor.lastrowid
-        
-        # Add sample goals
-        cursor = conn.execute('''
-            INSERT INTO goals (student_id, description, target_accuracy)
-            VALUES (?, ?, ?)
-        ''', (student_id, 'Improve articulation of /r/ sound in all positions', 85))
-        goal1_id = cursor.lastrowid
-        
-        cursor = conn.execute('''
-            INSERT INTO goals (student_id, description, target_accuracy)
-            VALUES (?, ?, ?)
-        ''', (student_id, 'Increase vocabulary comprehension and usage', 80))
-        goal2_id = cursor.lastrowid
-        
-        # Add sample objectives for goal 1
-        objectives_goal1 = [
-            'Produce /r/ sound in initial position with 85% accuracy',
-            'Produce /r/ sound in medial position with 85% accuracy',
-            'Produce /r/ sound in final position with 85% accuracy',
-            'Use /r/ words in connected speech with 80% accuracy'
-        ]
-        
-        for obj_desc in objectives_goal1:
-            conn.execute('''
-                INSERT INTO objectives (goal_id, description, target_percentage)
-                VALUES (?, ?, ?)
-            ''', (goal1_id, obj_desc, 85))
-        
-        # Add sample objectives for goal 2
-        objectives_goal2 = [
-            'Define new vocabulary words with 80% accuracy',
-            'Use new vocabulary in sentences with 75% accuracy',
-            'Answer comprehension questions about vocabulary with 85% accuracy'
-        ]
-        
-        for obj_desc in objectives_goal2:
-            conn.execute('''
-                INSERT INTO objectives (goal_id, description, target_percentage)
-                VALUES (?, ?, ?)
-            ''', (goal2_id, obj_desc, 80))
-        
-        conn.commit()
-        print("Sample data added successfully!")
-
-if __name__ == "__main__":
-    init_db()
-    add_sample_data()
